--- scripts/add/set-id.ts
@@ -1,12 +1,0 @@
-import { ObjectId } from "./settings";
-
-console.log(ObjectId());
-
-console.log(ObjectId());
-console.log(ObjectId());
-console.log(ObjectId());
-console.log(ObjectId());
-
-console.log(ObjectId());
-
-console.log(ObjectId());
