+++ src/common/pages/tournament/view/index.tsx
@@ -4,7 +4,7 @@ import { Tag } from "@blueprintjs/core";
 import * as React from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router";
-import Tournament from "../../../../server/models/tournament";
+import Tournament from "../../../..//models/tournament";
 import { IAppState } from "../../../reducers";
 const style = require("../view-all/style.css");
