+++ src/server/api/routes/team/post.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
-import Team from "../../../models/team";
-import User from "../../../models/user";
+import Team from "../../../../models/team";
+import User from "../../../../models/user";
 
 export default function(req: Request, res: Response) {
   const { name, location } = req.body;
