+++ src/server/api/routes/user/get.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import User from "../../../models/user";
+import User from "../../../../models/user";
 
 export default function(req: Request, res: Response) {
   const { expand = [], ...query } = req.query;
