--- src/server/models/division.ts
@@ -1,46 +1,0 @@
-import { Document, Schema } from "mongoose";
-import { model } from "./db";
-import Team from "./team";
-import Tournament from "./tournament";
-
-export enum Genders {
-  Boys = "Boys",
-  Girls = "Girls",
-  Mens = "Mens",
-  Womens = "Womens",
-  Open = "Open",
-  Mixed = "Mixed"
-}
-
-export interface DivisionSchema {
-  gender: Genders;
-  name: string;
-  teams: Team[];
-  tournament: Tournament;
-}
-
-export default interface Division extends DivisionSchema, Document {};
-
-export const DivisionSchema = new Schema({
-  gender: {
-    enum: Object.values(Genders),
-    required: true,
-    type: String
-  },
-  name: {
-    required: true,
-    type: String
-  },
-  teams: [
-    {
-      ref: "Team",
-      type: Schema.Types.ObjectId
-    }
-  ],
-  tournament: {
-    ref: "Tournament",
-    type: Schema.Types.ObjectId
-  }
-});
-
-export default model<Division>("Division", DivisionSchema);
